telegram_bot_response/requirements.txt
python-telegram-bot==21.1.1
requests
elasticsearch==8.13.2
python-dotenv